Added Dividers and Attempted Fix to CSS
Add dividers in the dropdown menu and and attempted to fix the issue of
"Event" disappearing after clicking it to close the dropdown menu. I
have since added code to make the menu appear on hover, but you can
still click event to toggle the menu and the issue is still there.